+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added fix for Flux Rifts not collapsing when fully consumed by Void Siphons
 - Added fix for Golems voiding held items when interacting with Use Seals with "Can use empty hand" enabled
 - Added tweak increasing vertical search range when the research table scans for nearby research aids
+- Added enhancement adding an effect over the player's health bar whenenever runic shielding is active
 
 ### Changed
 ### Fixed